labs/dvc_pipeline_lab/requirements.txt
# DVC Pipeline Lab Requirements

# DVC with S3 support
dvc[s3]==3.48.4

# Data handling
pandas==2.1.4
numpy==1.26.3

# Stock data download
yfinance==0.2.35

# Technical indicators
ta==0.11.0

# Data visualization (optional, for analysis)
matplotlib==3.8.2
seaborn==0.13.1

# AWS SDK for S3
boto3==1.34.34

# YAML parsing
PyYAML==6.0.1

# Progress bars
tqdm==4.66.1

# Scientific computing
scipy==1.11.4
